+++ misc/example-tester.bash
@@ -2,17 +2,40 @@
 
 set -eEuo pipefail
 
+if ! ps > /dev/null 2>&1; then
+  if yum --help > /dev/null; then
+    yum install procps -y
+  fi
+fi
+
+pidtree() (
+  set +u
+  [ -n "${ZSH_VERSION:-}"  ] && setopt shwordsplit
+  declare -A CHILDS
+  while read P PP;do
+    CHILDS[$PP]+=" $P"
+  done < <(ps -e -o pid= -o ppid=)
+
+  walk() {
+    for i in ${CHILDS[$1]};do
+      echo $i
+      walk $i
+    done
+  }
+
+  walk $1
+  set -u
+)
+
 function cleanup {
   set +e
   while true; do
-    PSAUX=$(ps fj | grep -v 'ps fj$' | grep -v "$0$" | grep -v grep | grep -v awk | awk '{print $2 " " $3}' | grep " $$$" | awk '{print $1}' | grep -v "$$")
-    PSAUX=$(printf "$PSAUX" | wc -l)
-    if [[ "$PSAUX" == "0" ]]; then
-      break
-    fi
-    pkill run_erl
-    pkill epmd
-    PIDS_TO_KILL="$(ps fj | grep -v 'ps fj$' | grep -v "$0$" | grep -v grep | grep -v awk | awk '{print $2 " " $3}' | grep " $$$" | awk '{print $1}' | grep -v "$$" | tr '\n' ' ')"
+    pkill run_erl > /dev/null 2>&1
+    pkill epmd > /dev/null 2>&1
+    pkill sender > /dev/null 2>&1
+    pkill data_receiver > /dev/null 2>&1
+    PIDS_TO_KILL="$(pidtree $$)"
+    PIDS_TO_KILL="$(ps -o pid= ${PIDS_TO_KILL})"
     if [[ "$PIDS_TO_KILL" == "" ]]; then
       break
     fi
@@ -21,7 +44,16 @@ function cleanup {
   set -e
 }
 
-trap cleanup SIGINT SIGTERM EXIT
+function error_cleanup {
+  local lineno=$1
+  local msg=$2
+  cleanup
+  echo "ERROR! Failed at $lineno: $msg"
+  exit 1
+}
+
+trap 'cleanup $LINENO "$BASH_COMMAND"' EXIT
+trap 'error_cleanup $LINENO "$BASH_COMMAND"' SIGINT SIGTERM
 
 LANG_TO_TEST=${1:-}
 
@@ -30,18 +62,22 @@ if [[ "$LANG_TO_TEST" == "" ]]; then
   exit 1
 fi
 
-HERE=$(dirname "$(readlink -f "${0}")")
-SOURCE_ACTIVATE="source $(readlink -f "${HERE}/../bin/activate")"
+HERE=$(dirname "$(readlink -e "${0}")")
+SOURCE_ACTIVATE="source $(readlink -e "${HERE}/../bin/activate")"
 BASH_HEADER="#!/bin/bash -ex"
 
-TESTING_TMP="${TMPDIR:-/tmp}/wally-up-testing"
+TESTING_TMP="${TMPDIR:-/tmp}/wallaroo-example-tester"
 
 rm -rf "$TESTING_TMP"
 mkdir -p "$TESTING_TMP"
 
-for dir in $(ls -d $(readlink -f "${HERE}/../examples")/${LANG_TO_TEST}/*/ | grep -v kafka | grep -v pony); do
+# TODO: Add logic to keep testing all examples even if one fails and give a summary/report of whichever ones failed at end.
+# TODO: Add ability to do the same for parts of the wallaroo book (install from source, vagrant, run an application, etc) where possible.
+for dir in $(ls -d $(readlink -e "${HERE}/../examples")/${LANG_TO_TEST}/*/ | grep -v kafka | grep -v pony); do
+  # call cleanup to kill any dangling processes started by this script to ensure a clean slate
+  cleanup
   CHANGE_DIRECTORY="cd $dir"
-  rm -f ${HERE}/../bin/metrics_ui/usr/var/log/*
+  rm -f ${HERE}/../tmp/log/*
   rm -f /tmp/$(basename "$dir")-*
   MY_TMP="${TESTING_TMP}/examples/${dir##*examples/}"
   mkdir -p "$MY_TMP"
@@ -55,6 +91,7 @@ for dir in $(ls -d $(readlink -f "${HERE}/../examples")/${LANG_TO_TEST}/*/ | gre
   echo "=================="
   echo "Working on example in directory: $dir"
   echo "=================="
+  echo "Scripts and log files can be found in: $MY_TMP"
   c=1
   while [[ $c -le 100 ]]
   do 
@@ -75,15 +112,17 @@ for dir in $(ls -d $(readlink -f "${HERE}/../examples")/${LANG_TO_TEST}/*/ | gre
   while [[ $d -lt $c ]]
   do
     if [[ "$d" -eq "($c - 1)" ]]; then
-      echo "Sleeping 60 seconds to let app run before shutdown..."
+      echo "Sleeping 60 seconds to let app run before shutdown (commands for last Shell)..."
       sleep 60
     fi
+    echo "Running script of commands for Shell ${d}..."
     LAST_CMD=$(tail -n 1 "$MY_TMP/shell${d}.bash")
     "$MY_TMP/shell${d}.bash" > "$MY_TMP/shell${d}.log" 2>&1 &
     last_pid=$!
     subshell_finished="false"
 
     # wait for bash to have executed the last command before proceeding
+    j=1
     while ! grep -F "+ $LAST_CMD" "$MY_TMP/shell${d}.log" > /dev/null 2>&1; 
     do
       # grep failed.. check to see if process is still running; if not, we might have a failure
@@ -93,8 +132,21 @@ for dir in $(ls -d $(readlink -f "${HERE}/../examples")/${LANG_TO_TEST}/*/ | gre
         wait ${last_pid}
         RET_CODE=$?
         if [[ "${RET_CODE}" != "0" ]]; then
-          echo "Error! Script for shell${d} failed with ${RET_CODE}!"
-          exit 1
+          if [[ "$LAST_CMD" != "metrics_reporter_ui start" ]]; then
+            echo "Error! Script for shell${d} failed with ${RET_CODE}!"
+            exit 1
+          fi
+          if [[ $j -gt 3 ]]; then
+            echo "Error! Script for shell${d} failed with ${RET_CODE}! Tried multiple times (because it is the metrics_ui) with no success!"
+            exit 1
+          fi
+          ## we have a failure for metrics_ui... try it again in case it was a weird transient thing..
+          cleanup
+          rm -f ${HERE}/../tmp/log/*
+          "$MY_TMP/shell${d}.bash" > "$MY_TMP/shell${d}.log.${j}" 2>&1 &
+          last_pid=$!
+          sleep 1
+          let j=j+1
         else
           # if zero, we didn't see the output we were expecting and we have a failure
           # don't fail immediately in case grep didn't get to see the full log file yet
@@ -118,8 +170,40 @@ for dir in $(ls -d $(readlink -f "${HERE}/../examples")/${LANG_TO_TEST}/*/ | gre
       ## check to make sure it's really started
       # sleep to avoid picking up first instance of start before restart/stop occur
       sleep 1
-      while [[ "$(tail -n 1 ${HERE}/../bin/metrics_ui/usr/var/log/erlang.log.1)" != "iex(metrics_reporter_ui@127.0.0.1)1> " ]]; do
+      i=1
+      while [[ "$(tail -n 1 ${HERE}/../tmp/log/erlang.log.1)" != "iex(metrics_reporter_ui@127.0.0.1)1> " ]]; do
         sleep 1
+        let i=i+1
+        wait $last_pid
+        RET_CODE=$?
+        if [[ "${RET_CODE}" != "0" ]]; then
+          if [[ $j -gt 3 ]]; then
+            echo "Error! Script for shell${d} failed with ${RET_CODE}! Tried multiple times (because it is the metrics_ui) with no success!"
+            exit 1
+          fi
+          ## we have a failure for metrics_ui... try it again in case it was a weird transient thing..
+          cleanup
+          rm -f ${HERE}/../tmp/log/*
+          "$MY_TMP/shell${d}.bash" > "$MY_TMP/shell${d}.log.${j}" 2>&1 &
+          last_pid=$!
+          sleep 1
+          let i=1
+          let j=j+1
+        fi
+        if [[ $i -gt 15 ]]; then
+          if [[ $j -gt 3 ]]; then
+            echo "Error! Metrics reporter taking too long to start! Tried multiple times (because it is the metrics_ui) with no success!"
+            exit 1
+          fi
+          ## we have a failure for metrics_ui... try it again in case it was a weird transient thing..
+          cleanup
+          rm -f ${HERE}/../tmp/log/*
+          "$MY_TMP/shell${d}.bash" > "$MY_TMP/shell${d}.log.${j}" 2>&1 &
+          last_pid=$!
+          sleep 1
+          let i=1
+          let j=j+1
+        fi
       done
     fi
 
@@ -138,30 +222,19 @@ for dir in $(ls -d $(readlink -f "${HERE}/../examples")/${LANG_TO_TEST}/*/ | gre
     if [[ "$d" -eq "($c - 1)" ]]; then
       echo "Waiting for shutdown commands to finish..."
       wait ${last_pid}
+      RET_CODE=$?
+      if [[ "${RET_CODE}" != "0" ]]; then
+        echo "Error! Script for shell${d} failed with ${RET_CODE}!"
+        exit 1
+      fi
     fi
 
     my_pids="$my_pids $last_pid"
     let d=d+1
   done
-  if [[ "$(pgrep -P $$)" != "" ]]; then
-    echo "Cleaning up errant child processes..."
-    PIDS_TO_KILL="$(ps fj | grep -v 'ps fj$' | grep -v "$0$" | grep -v grep | grep -v awk | awk '{print $2 " " $3}' | grep " $$$" | awk '{print $1}' | grep -v "$$" | tr '\n' ' ')"
-    # send sigterm to have them exit nicely
-    kill -15 $PIDS_TO_KILL
-    # wait for sigterm to do its thing and for child processes to exit cleanly
-    sleep 1
-
-    # if still not all exited
-    if [[ "$(pgrep -P $$)" != "" ]]; then
-      # TODO: should this be an error condition??
-      echo "Killing stubborn child processes..."
-      kill -9 $PIDS_TO_KILL
-      if [[ "$(pgrep -P $$)" != "" ]]; then
-        echo "WARNING: Not all processes ended between examples!!!"
-        ps fj | grep "$$"
-      fi
-    fi
-  fi
+  echo "Everything ran successfully. Cleaning up any errant processes..."
+  # call cleanup to kill any dangling processes started by this script to ensure a clean slate
+  cleanup
   echo "=================="
   echo "Done with example in directory: $dir"
   echo "=================="