+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to Wallaroo will be documented in this file.
 
 ### Added
 
+- Added Wallaroo Up to automate development environment setup on multiple Linux distributions
+- Added support for Fedora 28, CentOS 7, and Debian Stretch Linux distributions via Wallaroo Up
+- Added Vagrant as an option for trying out the Wallaroo Go
+- Added Docker as an option for trying out the Wallaroo Go
 
 ### Changed